+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 1.0.0 (February 27, 2018)
+- Breaking change: autolaunch.config is deprecated. Now simply add `"auto": true` to automatically run/launch the task/configuration.
+- Added schema validation for the `auto` property
+
 ### 0.0.3 (November 2, 2017)
 - Updated package.json
 - Added Github section to readme